# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Talai middleware",
    "description": "The only component allowed to talk to TallyPrime. Every route under /api/v1 requires a bearer token; /health does not.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/vouchers": {
      "get": {
        "tags": [
          "vouchers"
        ],
        "summary": "Vouchers from the replica",
        "operationId": "vouchers_api_v1_vouchers_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "name": "party",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Party"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vouchers/{voucher_id}": {
      "get": {
        "tags": [
          "vouchers"
        ],
        "summary": "One voucher",
        "operationId": "voucher_api_v1_vouchers__voucher_id__get",
        "parameters": [
          {
            "name": "voucher_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Voucher Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoucherDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bills": {
      "get": {
        "tags": [
          "vouchers"
        ],
        "summary": "Open bills with aging buckets",
        "operationId": "bills_api_v1_bills_get",
        "parameters": [
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(payable|receivable)$",
              "default": "payable",
              "title": "Direction"
            }
          },
          {
            "name": "as_on",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As On"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgingBucket": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "label",
          "amount",
          "count"
        ],
        "title": "AgingBucket"
      },
...
      "BillList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillOut"
            },
            "type": "array",
            "title": "Items"
          },
          "buckets": {
            "items": {
              "$ref": "#/components/schemas/AgingBucket"
            },
            "type": "array",
            "title": "Buckets"
          },
          "total_pending": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Pending"
          }
        },
        "type": "object",
        "required": [
          "items",
          "buckets",
          "total_pending"
        ],
        "title": "BillList"
      },
      "BillOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "party_ledger": {
            "type": "string",
            "title": "Party Ledger"
          },
          "bill_name": {
            "type": "string",
            "title": "Bill Name"
          },
          "bill_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bill Date"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "opening_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Opening Amount"
          },
          "pending_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Pending Amount"
          },
          "direction": {
            "type": "string",
            "title": "Direction"
          }
        },
        "type": "object",
        "required": [
          "id",
          "party_ledger",
          "bill_name",
          "bill_date",
          "due_date",
          "opening_amount",
          "pending_amount",
          "direction"
        ],
        "title": "BillOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VoucherDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "voucher_number": {
            "type": "string",
            "title": "Voucher Number"
          },
          "voucher_type": {
            "type": "string",
            "title": "Voucher Type"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "party_ledger": {
            "type": "string",
            "title": "Party Ledger"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "narration": {
            "type": "string",
            "title": "Narration"
          },
          "is_cancelled": {
            "type": "boolean",
            "title": "Is Cancelled"
          },
          "ledger_entries": {
            "items": {
              "$ref": "#/components/schemas/VoucherLedgerEntryOut"
            },
            "type": "array",
            "title": "Ledger Entries"
          },
          "inventory_entries": {
            "items": {
              "$ref": "#/components/schemas/VoucherInventoryEntryOut"
            },
            "type": "array",
            "title": "Inventory Entries"
          }
        },
        "type": "object",
        "required": [
          "id",
          "voucher_number",
          "voucher_type",
          "date",
          "party_ledger",
          "amount",
          "reference",
          "narration",
          "is_cancelled"
        ],
        "title": "VoucherDetail"
      },
      "VoucherInventoryEntryOut": {
        "properties": {
          "stock_item": {
            "type": "string",
            "title": "Stock Item"
          },
          "godown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Godown"
          },
          "qty": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "rate": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "hsn": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hsn"
          }
        },
        "type": "object",
        "required": [
          "stock_item"
        ],
        "title": "VoucherInventoryEntryOut"
      },
      "VoucherLedgerEntryOut": {
        "properties": {
          "ledger_name": {
            "type": "string",
            "title": "Ledger Name"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "is_deemed_positive": {
            "type": "boolean",
            "title": "Is Deemed Positive"
          },
          "cost_centre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Centre"
          }
        },
        "type": "object",
        "required": [
          "ledger_name",
          "amount",
          "is_deemed_positive"
        ],
        "title": "VoucherLedgerEntryOut"
      },
      "VoucherList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VoucherSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "VoucherList"
      },
      "VoucherSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "voucher_number": {
            "type": "string",
            "title": "Voucher Number"
          },
          "voucher_type": {
            "type": "string",
            "title": "Voucher Type"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "party_ledger": {
            "type": "string",
            "title": "Party Ledger"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "narration": {
            "type": "string",
            "title": "Narration"
          },
          "is_cancelled": {
            "type": "boolean",
            "title": "Is Cancelled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "voucher_number",
          "voucher_type",
          "date",
          "party_ledger",
          "amount",
          "reference",
          "narration",
          "is_cancelled"
        ],
        "title": "VoucherSummary"
      }
    }
  }
}